Added the taoster notification and confimation popup.

+++ src/app/components/all-policies/all-policies.component.ts
@@ -1,6 +1,9 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, TemplateRef } from '@angular/core';
+import { ToastrService } from 'ngx-toastr';
 import { policy } from 'src/app/models/policy';
 import { PoliciesService } from 'src/app/services/policies-service.service';
+import { ModalDismissReasons, NgbModal  } from "@ng-bootstrap/ng-bootstrap";
+//import { BsModalService, BsModalRef } from 'ngx-bootstrap/modal'
 
 @Component({
   selector: 'app-all-policies',
@@ -9,7 +12,8 @@ import { PoliciesService } from 'src/app/services/policies-service.service';
 })
 export class AllPoliciesComponent implements OnInit {
   policies = new Array<policy>;
-  constructor(private policyservice: PoliciesService) { }
+  constructor(private policyservice: PoliciesService,private toastr: ToastrService,
+    private modalService: NgbModal) { }
 
   ngOnInit(): void {
     this.policyservice.GetAllPolicy()
@@ -18,14 +22,21 @@ export class AllPoliciesComponent implements OnInit {
           this.policies = response;
         },
       })
-  }
-  OnDeleteClick(pid: any): void {    
-    this.policyservice.DeletePolicy(parseInt(pid))
+  } 
+ 
+  OnDeleteClick(content : any,pid: any) {
+    this.modalService.open(content, { ariaLabelledBy: 'modal-basic-title' }).result.then((result) => {  
+      if (result === 'yes') {  
+        this.policyservice.DeletePolicy(parseInt(pid))
       .subscribe({
         next: (response) => {
           this.policies = response;
+          this.toastr.warning("Deleted Successfully.")
         }
-      })
+      }) 
+      }  
+    }, (reason) => {  
+    });  
   }
 
 }
